Allow getting a new client with a context.Context
Context is great to allow for cancellation of requests.

With this change it will be possible to get a new client which will set
a different context for all it's requests.

This solution is backwards-compatible. However, since this package
always embeds itself to provide methods like *Board.GetLists(). The
current solution could lead to memory leaks or using old contexts if the
user is holding the responses in memory.